+++ Avisynth3/trunk/avisynth/src/filters/edit/splice/novideo.cpp
@@ -32,7 +32,7 @@ namespace avs { namespace filters { namespace splice {
 
 
 
-CPVideoFrame NoVideo::GetFrame(int /*n*/) const
+CPVideoFrame NoVideo::GetFrame(long /*n*/) const
 {
   throw exception::NoVideo();
 }
