+++ Avisynth3/trunk/avisynth/src/filters/edit/splice/novideo.h
@@ -54,8 +54,8 @@ public:  //structors
 
 public:  //clip general interface
 
-  virtual CPVideoFrame GetFrame(int n) const;
-  virtual BYTE * GetAudio(BYTE * buffer, long long start, int count) const { return audio_.GetAudio(buffer, start, count, *this); }
+  virtual CPVideoFrame GetFrame(long n) const;
+  virtual BYTE * GetAudio(BYTE * buffer, long long start, long count) const { return audio_.GetAudio(buffer, start, count, *this); }
 
 
 public:  //Splice interface
